+++ R/SMBassWB1.R
@@ -1,0 +1,18 @@
+#' @title Fish-specific data for West Bearskin Lake Smallmouth Bass.
+#'
+#' @description Fish-specific data for West Bearskin Lake Smallmouth Bass.
+#'
+#' @format A data frame with 445 rows and 10 variables:
+#' \describe{
+#'   \item{id}{An identification number unique to each fish.}
+#'   \item{species}{Fish species (all are \dQuote{SMB}).}
+#'   \item{lake}{Lake (all are \dQuote{WB}).}
+#'   \item{gear}{Capture gear (\code{E}=electrofishing, \code{T}=trapnet).}
+#'   \item{yearcap}{Year of capture.}
+#'   \item{lencap}{Total length-at-capture (mm)}
+#' }
+#' 
+#' @source Originally from XXX in \pkg{alr4}. Rearranged to demonstrate functions in this package.
+#' 
+#' @seealso \code{\link{SMBassWB2}}
+"SMBassWB1"