+++ tests/testthat/test_MESSAGES.R
@@ -69,6 +69,8 @@ test_that("digitizeRadii() error messages",{
 test_that("combineData() messages",{
   expect_error(combineData("small_ex.jpg"),"not an RData file")
   expect_error(combineData("notRFishBC.rds"),"does not appear to be from")
+  expect_error(combineData("Scale_1_DHO.rds",outFormat="Derek"),
+               "should be one of")
 })
 
 
@@ -112,16 +114,6 @@ test_that("bcFuns() messages",{
   expect_error(bcFuns("Derek"),"must be one of")
   expect_error(bcFuns("TVG"),"not yet implemented")
   expect_error(bcFuns(5),"not yet implemented")
-
-  ## List all choices for bcFuns() ( TVG is not included because it
-  ## is not yet implemented)
-  tmp <- c("DALE","FRALE","BI","LBI","BPH","LBPH","TVG","SPH","LSPH",
-           "AE","AESPH","AEBPH","MONA","MONA-BPH","MONA-SPH","WAKU",
-           "FRY","MF","ABI","FRY-BPH","ABPH","FRY-SPH","ASPH","QBPH",
-           "QSPH","PBPH","PSPH","EBPH","ESPH")
-  tmp <- tmp[-7]
-  ## Do all choices return a message with the name of the function in it
-  for (i in tmp) expect_message(bcFuns(i,verbose=TRUE),i)
 })